backend/requirements.txt
numpy==1.23.5
opencv-python==4.7.0.72
tensorflow==2.12.0
imutils==0.5.4
joblib==1.2.0
scikit-learn==1.2.2
pillow==9.5.0
pandas==2.0.1

fastapi==0.85.0
uvicorn==0.19.0

python-dotenv==0.21.0
